최종 병합 완료

+++ src/main/java/com/lms/domain/mafia/service/MafiaService.java
@@ -45,8 +45,9 @@ public class MafiaService {
             throw new RuntimeException(e);
         }
     }
+
     // 마피아 뽑기
-    public MafiaDTO selectMafia(int villageId) {
+    public MafiaDTO selectMafia(long villageId) {
 
 
         try {
@@ -68,7 +69,7 @@ public class MafiaService {
             MafiaDTO selectedMafia = list.get(randomIndex);
             // 다음 mafia_id 순번 계산 후 DTO에 세팅
             long nextId = mafiaDAO.selectNextMafiaId();  // 다음 순번 조회
-            selectedMafia.setMafiaId((long) nextId);           // DTO에 세팅
+            selectedMafia.setMafiaId(nextId);           // DTO에 세팅
             // 오늘 생성된 마피아 아이디
             selectedMafia.setCreatedAt(LocalDate.now());
             // DB에 넣겠디
